Monsters can use magic flutes
Taken from SpliceHack, but with changes. Other monsters in the area will
also be put to sleep, not just the hero (the flutist is immune); the
sleep time for the hero is also the same as that of
put_monsters_to_sleep. Like SpliceHack, they will not play the flute if
you are already immobilized somehow.

As it stands, monsters will never generate with magic flutes nor will
any ever pick one up due to none of them actually wanting to pick up
TOOL_CLASS items. More on this in the next commit.

Externify put_monsters_to_sleep and add a parameter for who is putting
the monsters to sleep (to center it on them and provide immunity).