Vampires can be staked through the heart
The list of criteria for this is kind of long:
1. The weapon hitting the vampire must be wood.
2. The weapon hitting the vampire must be a piercing weapon.
3. The vampire has to be in its normal form, not shapeshifted.
4. The vampire can't be wearing any body armor.
5. The hit has to qualify as a 'non-minimal' hit that will train skill.
6. You can't be twoweaponing.
7. It must be a critical hit (so 1/20 of otherwise qualifying hits).

If all these check out, you stake it through the heart and kill it
instantly. This deliberately applies to all V, including Vlad.

I did not add mhitm or mhitu cases for this; the reasoning is that it's
in the place in the code where there are other player-only special
attacks like shattering an enemy's weapon. I don't think this would
really ever happen that much in monster hands anyway. The flavor can be
handwaved as the savvy character knowing to aim for the heart.