Livelog when the player changes alignment
Permaconversion may not actually change your current alignment due to
wearing a helm of opposite alignment at the time, but it will still
livelog anyway. The messages between permaconversion and helm conversion
are different.

To avoid spam, I suppressed the livelog for removing a helm of opposite
alignment and turning back to the base alignment.