HTML Dumplogs - RC0
Initial HTML Dumplog implementation.
This extends the dumplog 'windowport' paradigm, and dumps text and
HTML dumplogs in parallel (though it is possible to configure just
one or the other via sysconf, or by enabling DUMPLOG/DUMPHTML
macros independently at compile time).

A minor consequence of this is that the status bar is now rendered
via the windowport, which may slightly change its appearance in the
text dump (eg: the hitpoint bar is rendered just as [title   ], and
enabling statuslines:3 in the game options will cause 3 status lines
to appear in both dump logs).

Original concept (and most of the CSS), from UnNetHack.