Fix: infinite loop with force_invmenu when not carrying anything
I don't remember whether this was a bug in the getobj patch and I was
waiting to see if it got fixed upstream in that patch, or not. Anyway,
this would trap you in an infinite output loop of "Not carrying
anything." Changing its continue to a break seems to fix the problem.