Fix: wizmode wishing for . (venom) caused rnd(0) impossible
Due to my earlier change that allows the probabilities of objects in a
class not to add up to 1000, if mkobj was called with the final valid
object class (which happens to be venom), it would incorrectly reference
bases[MAXOCLASSES].

Since the game never needs to make random types of venom, this only
happened during a wizmode wish for . but it was apparently triggering
the debug fuzzer quite a lot.

This augments mkobj to guard against that case, since it is within the
barest realm of possibility that a wizard mode user might actually want
to wish for a randomly chosen object from VENOM_CLASS. It wasn't
necessary to change any of the readobjnam code.