Fix: engraving with a wand of wishing gave a "free wish"
An oversight on my part when I first changed wands of wishing to
disintegrate on hitting 0 charges; that was only added for the zap
codepath and not for the engraving one. Thus, engraving with a 0:1 wand
would bring it down to 0:0 like normal, which could then be wrested by
engraving more.

This makes engraving also destroy the wand when it drops to 0:0, as was
intended originally. I didn't do anything for the breaking-the-wand
case, since that will obviously only grant you at most one wish.