Remove limit of 5 unrecognized wish attempts
This was an annoying quirk of the wishing code. If it couldn't parse
your input 5 times, it would give you a random object you almost
certainly don't want. What exactly is the point of this?

Fix this by allowing the user to enter malformed or unrecognizable input
an indefinite amount of times. They'll get it right eventually, or at
least get prompted to enter "help" for assistance and see their option
of getting out by entering "nothing", if they don't want anything.