Lots of new epitaphs, reorganize epitaph.txt
Some new epitaphs are original, a couple are from SpliceHack, and the
vast majority are from the nethack.alt.org user suggestion list. In
addition, I deleted some of the existing epitaphs that didn't feel quite
right coming from a gravestone. A couple are going to be turned into
rumors instead. Also made minor stylistic edits to both old and new
epitaphs. A few new epitaphs take advantage of the new ability to insert
things like the player's name.

Also put the long list of miscellaneous unclassified epitaphs (the bulk
of the file) into sorted order. Previously it was at least two kind of
sorted hunks with many exceptions.

Usually I update all the flavor data like these and engrave.txt, etc, in
the same commit, but this was enough of an overhaul to this one file
that I'm probably going to do each piece separately.