Pass get_rnd_text lines through the quest text parser
This allows for the addition of some interesting new lines in the
various data files (e.g. bogusmons, graffiti, etc) that are not static,
but instead will substitute various character-specific strings for the
format specifiers found in the quest text, such as %p, %r, and %a.

convert_line is externified so that it can be called from get_rnd_text,
and surprisingly only one rumor contained a % which needed to be
modified to %% so it'll turn into a literal % in the game.

(Rumors don't use get_rnd_text, they use get_rumor, but I added
convert_line into there as well.)