Lots of new true rumors
Most of these are from the nethack.alt.org suggestion list, but as
with the false rumors some are original or xNetHack specific.

For some of the preexisting ones that were ambiguous, not actually true,
or misleading, I deleted them or edited them to be clearer or more
accurate. (I'm being a little stricter on the true rumors file in
preparation for an upcoming feature: rumors.tru should contain only
*helpful* information, so its lines should be neither inscrutable nor
true only in very specific contexts.)