Lots of new bogusmons
There really are a lot of these this time. Most from nethack.alt.org but
a good number are original; some were even made up on the fly in the
process of merging all these in.

The bogusmons.txt file is one of the most broken-up due to the vast
number of things it references; I didn't significantly change the
overall layout of the file. Also not that much editing or removal of
existing bogusmons to be done because this file is a lot more "anything
goes".