New system for "cookie-only" rumors
There's apparently a system in place that will prevent certain rumors
from being chosen if they come from something other than a cookie. (So
that there can be false rumors like "You swallow the paper!" that won't
come from the Oracle or be found as graffiti.) The problem is that this
just did string comparison to treat every rumor containing "cookie" and
"pity" as cookie-only. This obviously isn't a flexible system.

The new system has the author of the rumors file insert a : at the start
of the line to mark it as cookie-only, which gets stripped out later if
that rumor is selected. I added an explanation of this at the top of
both rumor source files as well, and put the : on the existing
cookie-only rumors.