New system for delivering angelic/demonic maledictions
The system as it existed previously was enormously inflexible - the
lines were stored in quest.txt, needed to be properly numbered (and
renumbered if any of them were removed), and had a small maximum amount
to fit into a small range of numbers (making it quite difficult to
increase the number of taunts, because that would require renumbering
all higher general-purpose text blocks).

The only reason I could think of for why they were stuck in quest.txt
was that some of the taunts needed to use % substitutions. But it's not
hard anymore to use convert_text and convert them for purposes other
than reading out of quest.txt.

So this commit tosses all that out, puts the maledictions in wizard.c
(near the cuss() function and the static strings used for the Wizard's
personal maledictions) and replaces the com_pager calls with verbalizing
those strings. I also added several new taunts to both, most from
nethack.alt.org's user suggestion list.

Quest nemesis maledictions and leader and guardian dialogue lines remain
in quest.txt.