Generate a true rumor graffito in the dungeon's starting room
This is my attempt at providing something similar to UnNetHack's free
hint scrolls, but in a way that delivers helpful in-game strategy advice
(UnNetHack hints all currently either give the player an interface tip
or refer them to seek help online). Graffiti are also less intrusive
than Un's guaranteed-until-configured-out mail daemon appearance, and
don't provide any information that couldn't generate elsewhere on the
same level.

The rumor is not eroded; this seemed like the best way to deliver the
message to an inexperienced player (and also occasionally deliver
xNetHack-specific rumors to experienced players).