Helmets made of hard nonmetal materials protect just as well as metal
From EvilHack, though I found several more places in the code where this
needed to be added. The gist is that a helm made of wood, bone, stone,
or glass will reduce damage taken from an object falling on your head as
if it were metal. EvilHack doesn't do this for glass helms, but they are
hard, so there you go.