Glass helms can break when heavy things fall on your head
A large blunt impact from a falling boulder or iron ball (not regular
rocks) might shatter a glass helmet. This is subject to the checks in
break_glass_obj, so the helmet will shatter 1/6 of the time unless it's
shatterproofed or an artifact. In such a case, the hardness of the helm
doesn't reduce the damage, though it will as normal for a hard helm if
it doesn't shatter.

Also add a special case: a glass piercer falling onto a glass helm will
always break it (unless shatterproofed or an artifact) and deal a little
extra damage.