Prevent monsters' starting gear from being materials they hate
Bug discovered in EvilHack based on xNetHack code, but fix is totally
different. Rather than enumerating all the possible failure cases and
hardcoding appropriate materials for each, just prevent mongets from
allowing the item it's creating to be a material the monster would hate.

Not covered by this fix are any items not created by mongets: the few
mksobj cases in makemon.c, most non-weapon non-armor gear. (That is, you
might still get a demon generated with a silver wand of striking. This
might actually be a vanilla bug.)

Also externify init_obj_material to make this possible.