Reduce the protective effect of a hard helm for heavy falling things
Let's be realistic: if a boulder or iron ball falls on your head, it's
going to do a lot of damage no matter how indestructible your helmet is.

This reduces the damage at most 2 points when the thing hitting your
head weighs more than 500 aum. So regular old rocks and most thrown
objects still get the "Fortunately" message as normal and usually reduce
damage to 2 points or so; iron balls, boulders, and thrown loadstones
barely notice whether you're wearing a helmet.

This also removes the behavior where a hard helm protects completely
against a falling piercer. Rather than not giving any damage at all, cut
it in half instead. The chance of dodging it based on AC is not changed.

Also tweak the > 500 cases for breaking glass helms to >= 500 so that
loadstones are able to break them as well.