Fix: only try to find alternate materials for worn/wielded gear
Bug discovered by K2 in which the game tried to give an iron wand to an
elf and since iron was the only valid material for that particular
object in that particular game, it was incapable of finding a suitable
material.

Vanilla doesn't currently seem to treat zapping a wand as touching that
object. Follow that lead for now: only weapons, armor, weapon-tools, and
amulets are forced to be a material that suits the monster.