Scale monster group size based on dungeon level, not character level
There are an lot of things in NetHack that scale on the player's
experience level. This was one of them. But the group size should depend
on the local difficulty of the area, not the player - this makes flavor
sense (why would a level 1 pacifist on dungeon level 23 encounter
smaller hordes of orcs than an appropriately leveled player would?) and
also practical sense (reducing your own level shouldn't be rewarded with
an easier dungeon.)

This also removes a compiler bug hack in m_initgrp that was last
reported at least sixteen years and 6 major GCC versions ago, and only
affected systems I don't really expect xNetHack to be compiled on. It's
probably not going to cause any problems.