Crystal ball revamp
Quite a few of crystal balls' mechanics were changed up in this commit.
The most significant is that they no longer make you helpless; looking
into them is now an occupation and can be interrupted by normal means.
Other changes include:
 * You are not prompted to look for a symbol unless you're actually
   going to get a vision of it. (Previously, you could get the prompt
   for an uncharged crystal ball).
 * If you enter an unrecognized character, you get prompted again until
   you enter a valid one or escape out of the prompt.
 * The outcomes where you would see a special level and your relative
   location to it (only caused previously by entering unrecognized
   characters) don't exist now, and the code for them is removed.
 * Cancelled crystal balls produce a special message and nothing special
   will happen with them. According to my testing, it's not even possible
   to cancel a crystal ball, but the code's there.
 * Uncharged crystal balls produce another unique message (gray fog).
 * The ball now takes 6-10 turns to look into, not 1-10.
 * Rather than a vague "vision is unclear", give a more explicit "You
   don't seem to find any of those here."
 * The Wizard of Yendor now might be found gazing out at you twice as
   often as he was previously.

You spend time looking into the crystal ball before you get the prompt
for what you want to look for (which is your opportunity to cancel out).
I considered prompting the user for the symbol before you started
peering in, but other tools don't ask you if you want to cancel them
after you apply them. Canceling out doesn't consume a charge, so the
only negative consequences are the turns lost and the fact that you will
need to pass the "oops" checks the next time you want to use it.

Due to the occupation requiring adding new fields to struct context,
this is a savebreaking change.