Fix: splitting tin stack gave it new label
Because labels depended deterministically on o_id, and when you split a
stack the split object gets a new one.

Fix this by overloading the 'usecount' field (currently only used for
spellbooks) to store a random number that gets set once when the tin is
created, and use this number instead of o_id. Splitting a stack will
keep that number the same.

I didn't add a condition that prevents two otherwise mergable tins from
merging if they had different labels. In the interest of gameplay, since
the tins are *functionally* identical, it's probably best to let them
merge.