Implement debug fuzzer enhancements from EvilHack
Adds a fuzzing log that stores basic details of what's happening in the
game so that when gdb captures the fuzzer when it panics you can see
what happened leading up to that. This was useful in fixing an
insane object error I was getting, where I could see how the object was
invalid but not how it got that way.