Don't generate random traps as falling rock on outdoors special levels
This led to the silly "A trap door in the sky opens and a rock falls
out!" Great that the game handled this smoothly, but this was just
ridiculous.

I still haven't added (and may never add) support for marking certain
areas of an outdoors map as indoors, so if a level is half outside and
half inside, you'll get rocks falling from the sky if it's not marked
outdoors, and you'll get no falling rock traps at all if it is marked
outdoors.

The only remaining place where a falling rock trap might generate on an
outdoors level is in a niche, and those aren't likely to appear on
outdoors levels.