Gas clouds: better randomization
First, clouds now perform their outward breadth-first-search with a
randomized list of directions rather than iterating in left, right, up,
down order. This removes the bias towards the upper left in cloud
spreading.

Then, give the cloud a 1/2 chance of not spreading to a space it
otherwise would have spread to. This has the effect of fuzzing the cloud
edges in open areas, helping the clouds on, for instance, the Plane of
Fire not be big rhombuses.

I later noticed a problem where if you centered the gas cloud in, say, a
corridor, it would probably fail the 1/2 chance of spreading in each
direction before going too far, meaning the cloud would end up tiny.
Addressed this issue by only allowing the search to skip on the fourth
valid spot it finds (so, this only happens in totally unconstrained
areas; the search progressing in, say, a corridor won't usually find
four valid spots.)