Object materials: extend to statues
Statues are usually still stone, but they can now be copper or gold.
Metallic statues are currently unbreakable. Trying to dig one manually
will still spend several turns on the attempt, but it has no effect.
They also cannot be stone-to-fleshed. However, it will animate normally
if it's the statue on a statue trap.

Statues now always show their material (what used to be "a statue of a
newt" is now "a stone statue of a newt").