Place Fort Ludios portal in the first eligible vault generated
From UnNetHack. No modification is made to the prerequisites for the
portal - a vault in the main dungeon below level 10, not on a level
where another branch splits off - but it will now not randomly decide to
skip the vault when generating a level and maybe defer to some later
vault.

There are two reasons for this. First, to make Ludios accessible in
practically every game. It will only be inaccessible if no vault
generates on every single level between level 11 and the Castle, which is
a very small chance. Second, because this biases the portal to be placed
rather high up in the dungeon, often before the hero's really ready to
take on Ludios, and it's more interesting to have a branch that is
discovered then deferred and returned to later than a branch that is
discovered then immediately cleared without much effort.