New ammo breakage formula
The bulk of this is taken from NetHack Fourk. The problem that Fourk
aimed to solve is that early-game characters who start with ammo
(Ranger, Tourist, Samurai) had their ammo break too quickly in the
normal course of combat, and were thus forced into relying on melee
weapons because ranged weapons are useless after the ammo is gone.

The main revision here is that your current and maximum skill levels in
the type of ammo you are using are the main factor in whether the ammo
breaks. Maximum skill is used to provide a boost to those early-game
ranged characters. Some examples, not accounting for enchantment:
 * Restricted: breaks 1/2 of the time
 * Unskilled, cap is Basic: breaks 1/4 of the time
 * Unskilled, cap is Skilled: breaks 1/6 of the time
 * Unskilled, cap is Expert: breaks 1/8 of the time
 * Basic, cap is Skilled: breaks 1/12 of the time
 * Basic, cap is Expert: breaks 1/16 of the time
 * Expert, cap is Expert: breaks 1/32 of the time

Then, enchantment and beatitude get a chance to change the outcome, with
beatitude acting as a temporary +2/-2 to enchantment. Positive
enchantment can save a breaking object, or negative enchantment can
destroy an intact object, with |ench|/(|ench|+1) chance. These rules are
different from Fourk's approach, which just gives cursed ammo an extra
Luck-based chance of breaking.