Increase the number of summoned demons for demon princes/lords
From EvilHack. Demon princes now gate in twice as many demons as a
rank-and-file demon will; lords will gate at least as many as a
rank-and-file demon, possibly randomly as many as a prince would. As
before, if another demon lord/prince gets selected to be gated in, no
other demons will be created.

Also change the "cloud of smoke" appearance message to "brimstone", also
from that EvilHack commit.