Goblins now generate in small groups
From EvilHack, which got it from SporkHack. Due to monster group size
scaling, this actually only affects goblins generated further down in
the dungeon - you'll only be able to see a group of 2 generate starting
at level 3 or so, a group of 3 isn't possible until level 5, and a group
of 4 not until level 7.