Fix: SPFX_WARN artifacts without a glow color specified shouldn't glow
Previously they glowed 'no color' in inventory, though special casing in
Sting_effects would mean that the 'glows!'/'stops glowing' messages
would not be shown.

This makes it so that any artifact that does have a glow color specified
will get the Sting_effects messages (previously it was hardcoded for
Sting, Orcrist and Grimtooth), and any artifact that doesn't will no
longer show up as glowing in the inventory, though it continues to warn
of whatever monster type it warns against like usual.

Side effect of this fix is that Demonbane now tells you when it
starts/stops glowing white.