Fix: most getobj prompts would ignore the specified count
I still don't quite get the logic of this function, but this one was my
fault. In the initial patch implementation, it would set an important
variable 'res' which would send the code into a new block that would
handle the count and split the stack properly. When silencing compiler
warnings, gcc was complaining about this function having a non-returning
path, so I changed the break out of the "find obj in inventory" loop to
merely return the object instead, which discarded the count entirely.

I tried replacing it with the original break statement and it compiles
without any warnings now, for some reason. Maybe a newer gcc version is
smart enough to follow the loop back around and see that res must get
set which prevents the function from having no return value.