Material damage for contact with monsters made of a hated material
This corrects the logical issue of why an iron golem isn't dealing extra
iron damage if you're an elf and it's punching you. Now that damage will
be dealt properly. Implemented this for both uhitm and mhitu cases. It
doesn't cover mhitm, since in the process of this I discovered that
mhitm actually never had code to account for monsters attacking each
other with hated materials in the first place. Once the DevTeam fixes
that, I'll implement the mhitm case of this.

Note that if you are punching elves as an iron golem but wearing gloves,
the gloves will fully override this (they'll still deal iron damage if
they are iron, but won't if they are something else.) Same for boots
with kicking.

Though the only monster made of a material that something else hates is
currently the iron golem, the code should generalize to any material
hatred introduced in the future.