Improve hunger messages and add a message for no longer hungry
More bettering of messages: "You are starting to feel weak" isn't
particularly clear, especially because of conflation of "weak" with
other things that drain strength.

Also add a new message when you eat food and become NOT_HUNGRY as a
result. Players that don't want to see this can MSGTYPE=hide it.