Large stone to flesh refactor: turning into meatballs
This started off as a fix for the weirdness in which stone to fleshing a
statue of a large, vegetarian monster (such as a gelatinous cube) would
produce a single meatball, and spiraled into becoming a big refactor.

In the process, I generalized this behavior for vegetarian monsters into
any case in which the statue or figurine does not animate for whatever
reason (previously it would either turn into a corpse, or just sit there
as if it weren't made of stone). I also removed a case I added earlier
for object materials that could never get reached.

The current behavior for any other type of object the function wasn't
specifically checking for is to see if it is valid to turn its material
to FLESH (nothing is currently valid as both STONE and FLESH). If this
is not valid, turn it into a number of meatballs (the object's weight
divided by 10).