Object lookup: show the relevant skill or launcher type for weapons
Skill is used unless it's ammo, in which case it prints the skill name
which happens to be the same as the launcher name.

Caveat, this does not describe rocks of any sort as "ammunition meant to
be fired from a sling", because this change only affects WEAPON_CLASS
objects.