Scrolls of scare monster do not affect unique monsters
A common criticism of the 3.6.0 Elbereth revamp is that it elevated
scrolls of scare monster to be an incredibly powerful item, since you
could drop them on the ground to prevent every single enemy in the game
from attacking you in melee (even minotaurs which were previously
immune), and you could freely attack while standing on them, unlike
Elbereth. Most significantly, it could make short work of quest nemeses
and demon lords, trivializing some of the hardest bosses in the game.

This is an approach that's been taken in the other post-3.6 variants:
unique monsters are not affected by this scroll anymore.