Fix: explicit xkill flags for burning through a slime engulfing you
This just used a mysterious 3 before, but we have explicit flags for
xkilled, so use them. Also, add in XKILL_NOCONDUCT: when you kill a
slime just by sitting there as a fiery monster, you should get credit
and experience for the kill, but you didn't actively kill it so don't
break pacifist conduct.