Trap ammo overhaul, part 1: implement framework
This sets up the essentials for the "trap ammo overhaul" - associating
an object with a trap so that it can persistently refer to that. No
changes to actual traps have happened yet - for instance, dart traps
still create darts out of thin air and have a random 1/15 chance of
disappearing after their first use.

Traps now have a new field "ammo", which holds the object. This commit
sets up the appropriate save and restore code and also defines a new
obj->where for being inside a trap.