Trap ammo overhaul, part 2: use ammo field for 4 trap types
Dart and arrow traps now store a stack of their respective ammo in the
ammo field. Repeated triggering will produce the same objects as would
come out by untrapping, and the same number. The amount in the stack for
both kinds of traps is boosted to 15 + rnd(20). I also merged much of
the dart and arrow trap logic, because it was very similar.

Bear traps and landmines now store a single beartrap or land mine
object in the trap.

Minor gameplay changes: darts now call the various thit() routines with
tlev = 8, making darts a little more likely to hit you; and poison darts
from a trap now deal Con damage 60% of the time and bonus HP damage 40%.

This also renames cnv_trap_obj to remove_trap_ammo, and refactors it to
remove its otyp and count parameters (which are now useless, as it
doesn't need to generate a new object on the fly.)