Poison dart traps won't appear until level difficulty 7 or higher
Because the effects of a poison dart got rougher, delay their
appearance for the earliest levels of the dungeon. Stepping off the DL1
upstairs and immediately losing 3 Con, or all your HP, to a poison dart
is not very fun. This also makes it a bit harder to get your hands on
poison darts early in the game.