Re-add Frontier Town as an eighth Minetown variant
I was never sure why SporkHack (and thus indirectly 3.6.0) removed
Frontier Town completely to replace it with Orcish Town. The level is a
perfectly fine Minetown variant as it is.

This also helps fix a problem several developers have noted, which is
that Orcish Town really needs to just not show up as often. With
uniform probabilities for all Minetown variants, this reduces the chance
from 1/7 to 1/8.