Add celibate conduct
Moderately popular unofficial conduct in vanilla, and can be used to
identify games where the player didn't rely on foocubi for anything.
Notably, this can be broken against the player's wishes (it's not really
possible to determine whether any given foocubus interaction was
intentional or not), but there are ways to guard against it.

I suppose some players might want this conduct to instead be "don't gain
any beneficial effects from a foocubus", but that's not as neatly
defined.