Make Caveman quest narration and dialogue caveman-like
The goal here is to make the quest feel more flavorful and distinct from
the rather samey text of the other quests. (There are some long term
plans in the works to change quest text, among other things, to vary up
the storylines - this isn't really part of that effort though. Among
other things, this aims to fix the weirdness of the fellow society of
primitive people all speaking very articulately.)

The narrator and the fellow cave-dwellers now speak in a stereotypical
primitive voice of mostly monosyllables. Not entirely monosyllables -
they can pronounce "Tiamat", after all, and I didn't want to change her
name to "Great Fierce Beast" like in Fourk.

Most of the new text is drawn directly from NetHack Fourk, lightly
edited on occasion.