Generalize loadstone behavior into new undroppable() macro
I plan to do something along the lines of making cursed gold items
undroppable, so I needed to change a lot of loadstone special cases into
non-loadstone cases while retaining behavior. Some loadstone-specific
things that are not related directly to being undroppable (such as being
able to be picked up with a full inventory, or disregarding weight
considerations) remain specific to loadstones.

This removes the corpsenm kludge in canletgo(), because we don't want to
clobber corpsenm on arbitrary undroppable objects. All this affected was
whether an "any of" appeared in the "For some reason, you cannot
drop..." message. Now it will simply print "any of" if there is more
than 1 undroppable object in the stack you are trying to let go of.

It also required a refactor of the oselect function, which is
responsible for choosing a weapon for a monster. Since loadstones can no
longer be a special case here, oselect had to be made aware of whether
it was being used to choose something that might be thrown or not.