Remove loadstones
This has been a long time coming. Loadstones are the prime example of an
object in NetHack that has absolutely nothing going for it, worse than
useless. There have been a few feeble proposals to make them worthwhile
somehow, but their enormous weight precludes that from ever happening.

The Mimic of the Mines' loadstone is replaced with a touchstone. Player
monsters will generate with a flint stone where they would previously
have received a loadstone.

This also deletes various behaviors peculiar to loadstones that I don't
intend to add back in for anything:
 - Autocursing when dropped
 - Automatically taking the whole stack off the ground or out of a
   container (it's still not possible to *drop* part of a stack of
   undroppables though)
 - Skipping the 52 item limit and weight considerations when being
   placed into the inventory
 - Being counted as the worst possible cursed item for prayer trouble
   purposes (undroppables still count towards this trouble, but they are
   the lowest priority behind all currently worn items)