Add messages for cursed worn or wielded items becoming uncursed
This is a nice indication that the curse on something that is currently
stuck to you got removed. Of course, you could have previously learned
this by checking your inventory, but it's good to provide a message.

Note that a message is only printed if the bknown flag was set - if you
were wearing cursed gear that you hadn't learned was cursed, you won't
notice anything when it becomes uncursed.