Update patchlevel for version 5.0
Note that this is version 5.0, which was previously referred to as
0.5.0. I decided that I'm never going to do anything so drastic to the
codebase that it would make me bump the major version 0 to a 1, and it
would certainly never happen multiple times. I don't really see a point
in using the third digit, either, because non-savebreaking changes are
generally rolled out as they are implemented, so using the first digit
for big changes and the second digit for smaller but still savebreaking
changes seems fine.