Make Ran-home floor normal instead of grass
I intended to add patches of grass in random places, but after
implementing the requisite lua code I realized that replace_terrain with
selections doesn't exist, which is... kind of needed for this. So, only
the first part of removing all the grass. (A totally green level wasn't
working out visually.)