Lua-ify Bar-goal grass gradient selection (still commented out)
I went through the process of transitioning this bit to Lua now that
gradient selections are available, but then found out you can't
replace_terrain with a selection (yet). In the hopes that that gets done
by the devteam at some point (since they have a TODO for it), I've left
it commented out.