Fix: couldn't wish for up to 5000 gold pieces
At some point the handling for gold in the wish parser has diverged from
vanilla. So gold ends up going through a check here that it doesn't in
vanilla, namely the usual "wishing for a stack" check that sets the
quantity of the received item to 1 if you wished for more than rnd(6) of
them.

Add an exception to this for COIN_CLASS wishes. (Execution will then hit
the check which caps gold at 5000 correctly.)