Revamp Juiblex's properties and behavior
This implements a patch from paxed which is on a WIP branch from 3.7 (so
it might end up getting added to 3.7), in which Juiblex can split into
multiple Juiblexes using the same logic as a pudding, and if you kill
the original Juiblex, a randomly selected clone becomes the new Juiblex.
Modifications I made on top of paxed's work: the cloned Juiblexes can
split as well, and the main non-cloned Juiblex does not lose any HP by
having a copy split off (a suggestion from bhaak since otherwise he
weakens pretty fast).

He is also made into a covetous non-warper (meaning he still covets the
Amulet and will attack adjacent things on or holding it but won't warp
to stairs). To compensate for this:
- his speed is doubled (which is still only 6, pretty slow)
- he may choose to split himself on his own initiative if there isn't
  already a clone; if there is, his bodies will produce lesser oozes and
  blobs
- he can hide on the ceiling, and starts in his level hidden on the
  ceiling

Note that there is some emergent behavior in making him a non-warper:
since he has a spit attack which is nearly or entirely unused in vanilla
because he will always covetous warp, he now falls into the general
logic for monsters with spit attacks and avoids closing to melee range
and will evade (though will still attack in melee if you are adjacent).
I didn't see a clean way to carve out an exception for this, and then
thought about it and decided it's not actually that bad, since like all
the other demon lords, the goal is now to defeat him, rather than just
pass through his level.

One adjustment that needed to be made in xNetHack is to copy the
mstrategy when a monster is cloned, and prevent the APPEARMSG flag from
being set, so that the Juiblex-clones don't trigger duplicate copies of
the Juiblex appearance message.